01: Clarify C's manual deallocation is for the heap
Manual deallocation is only required in C for memory allocated on the
heap; the stack is automatically freed even with alloca/VLAs. Insert the
word "heap" here to clarify this ("heap" is used later in the lecture so
I assume it's a familiar concept for students.)

Even this description is a bit unfair to C... modern C projects also
avoid manual deallocation with techniques like managed allocations
(Linux kernel, Mesa) and streamlined reference counting (glib). However,
these techniques are not part of the language like garbage collection is
part of Java, so I think this is the right wording.

Signed-off-by: Alyssa Rosenzweig <alyssa@rosenzweig.io>